+++ erpnext/fixtures/custom_field.json
@@ -1,0 +1,56 @@
+[
+  {
+    "allow_in_quick_entry": 0,
+    "allow_on_submit": 0,
+    "bold": 0,
+    "collapsible": 0,
+    "collapsible_depends_on": null,
+    "columns": 0,
+    "default": null,
+    "depends_on": null,
+    "description": null,
+    "docstatus": 0,
+    "doctype": "Custom Field",
+    "dt": "User",
+    "fetch_from": null,
+    "fetch_if_empty": 0,
+    "fieldname": "department",
+    "fieldtype": "Link",
+    "hidden": 0,
+    "hide_border": 0,
+    "hide_days": 0,
+    "hide_seconds": 0,
+    "ignore_user_permissions": 0,
+    "ignore_xss_filter": 0,
+    "in_global_search": 0,
+    "in_list_view": 0,
+    "in_preview": 0,
+    "in_standard_filter": 0,
+    "insert_after": "username",
+    "is_system_generated": 0,
+    "is_virtual": 0,
+    "label": "Department",
+    "length": 0,
+    "mandatory_depends_on": null,
+    "modified": "2023-03-21 10:23:59.353469",
+    "module": "ERPNEXT",
+    "name": "User-department",
+    "no_copy": 0,
+    "non_negative": 0,
+    "options": "Department",
+    "permlevel": 0,
+    "precision": "",
+    "print_hide": 0,
+    "print_hide_if_no_value": 0,
+    "print_width": null,
+    "read_only": 0,
+    "read_only_depends_on": null,
+    "report_hide": 0,
+    "reqd": 0,
+    "search_index": 0,
+    "sort_options": 0,
+    "translatable": 0,
+    "unique": 0,
+    "width": null
+  }
+]
