+++ erpnext/fixtures/translation.json
@@ -1,0 +1,15 @@
+[
+ {
+  "context": null,
+  "contributed": 0,
+  "contribution_docname": null,
+  "contribution_status": "",
+  "docstatus": 0,
+  "doctype": "Translation",
+  "language": "en",
+  "modified": "2024-01-04 14:55:52.942013",
+  "name": "7456b18947",
+  "source_text": "Material Request",
+  "translated_text": "Internal Requisition"
+ }
+]
