--- erpnext/fns/doctype/internal_requisition/internal_requisition.json
@@ -1,145 +1,0 @@
-{
- "actions": [],
- "allow_rename": 1,
- "autoname": "naming_series:",
- "creation": "2023-12-10 02:09:08.140853",
- "default_view": "List",
- "doctype": "DocType",
- "editable_grid": 1,
- "engine": "InnoDB",
- "field_order": [
-  "irf",
-  "naming_series",
-  "transaction_date",
-  "t_warehouse",
-  "colum_break1",
-  "schedule_date",
-  "warehouse_section",
-  "scan_barcode",
-  "internal_requisition_items",
-  "clb1",
-  "faculty_button",
-  "faculty_items",
-  "amended_from"
- ],
- "fields": [
-  {
-   "fieldname": "transaction_date",
-   "fieldtype": "Date",
-   "label": "Transaction Date",
-   "reqd": 1
-  },
-  {
-   "fieldname": "colum_break1",
-   "fieldtype": "Column Break"
-  },
-  {
-   "fieldname": "t_warehouse",
-   "fieldtype": "Link",
-   "in_list_view": 1,
-   "label": "Target Warehouse",
-   "options": "Warehouse",
-   "reqd": 1
-  },
-  {
-   "allow_on_submit": 1,
-   "fieldname": "schedule_date",
-   "fieldtype": "Date",
-   "label": "Required By"
-  },
-  {
-   "fieldname": "warehouse_section",
-   "fieldtype": "Section Break",
-   "hide_border": 1,
-   "label": "Items"
-  },
-  {
-   "allow_on_submit": 1,
-   "fieldname": "scan_barcode",
-   "fieldtype": "Data",
-   "label": "Scan Barcode",
-   "options": "Barcode"
-  },
-  {
-   "fieldname": "amended_from",
-   "fieldtype": "Link",
-   "label": "Amended From",
-   "no_copy": 1,
-   "options": "Internal Requisition",
-   "print_hide": 1,
-   "read_only": 1,
-   "search_index": 1
-  },
-  {
-   "fieldname": "irf",
-   "fieldtype": "Section Break",
-   "label": "<h4><b>Internal Requisition Form<b></b></b></h4>"
-  },
-  {
-   "fieldname": "naming_series",
-   "fieldtype": "Select",
-   "hidden": 1,
-   "label": "Series",
-   "no_copy": 1,
-   "options": "REQ-IR-.YYYY.-",
-   "print_hide": 1,
-   "set_only_once": 1
-  },
-  {
-   "allow_on_submit": 1,
-   "depends_on": "eval:doc.workflow_state==\"Endorsed\"",
-   "description": "Please Issue quantity.",
-   "fieldname": "clb1",
-   "fieldtype": "Tab Break",
-   "label": "Faculty Office Section"
-  },
-  {
-   "fieldname": "internal_requisition_items",
-   "fieldtype": "Table",
-   "label": "List of Items",
-   "no_copy": 1,
-   "options": "Internal Requisition items"
-  },
-  {
-   "fieldname": "faculty_items",
-   "fieldtype": "Table",
-   "label": "Quantity",
-   "options": "Faculty Items"
-  },
-  {
-   "depends_on": "eval:doc.workflow_state=='Endorsed'",
-   "fieldname": "faculty_button",
-   "fieldtype": "Button",
-   "label": "Click to Issue Items"
-  }
- ],
- "index_web_pages_for_search": 1,
- "is_submittable": 1,
- "links": [],
- "modified": "2024-01-02 14:44:46.250345",
- "modified_by": "Administrator",
- "module": "FNS",
- "name": "Internal Requisition",
- "naming_rule": "By \"Naming Series\" field",
- "owner": "Administrator",
- "permissions": [
-  {
-   "create": 1,
-   "delete": 1,
-   "email": 1,
-   "export": 1,
-   "print": 1,
-   "read": 1,
-   "report": 1,
-   "role": "System Manager",
-   "share": 1,
-   "write": 1
-  }
- ],
- "sort_field": "modified",
- "sort_order": "DESC",
- "states": [],
- "track_changes": 1,
- "track_seen": 1,
- "track_views": 1
-}
