--- erpnext/fns/doctype/internal_requisition/internal_requisition.py
@@ -1,8 +1,0 @@
-# Copyright (c) 2023, Frappe Technologies Pvt. Ltd. and contributors
-# For license information, please see license.txt
-
-# import frappe
-from frappe.model.document import Document
-
-class InternalRequisition(Document):
-	pass
