--- erpnext/fns/doctype/internal_requisition/test_internal_requisition.py
@@ -1,9 +1,0 @@
-# Copyright (c) 2023, Frappe Technologies Pvt. Ltd. and Contributors
-# See license.txt
-
-# import frappe
-from frappe.tests.utils import FrappeTestCase
-
-
-class TestInternalRequisition(FrappeTestCase):
-	pass
