+++ erpnext/stock/doctype/material_request_item/material_request_item.json
@@ -83,6 +83,7 @@
    "fieldname": "item_name",
    "fieldtype": "Data",
    "in_global_search": 1,
+   "in_list_view": 1,
    "label": "Item Name",
    "oldfieldname": "item_name",
    "oldfieldtype": "Data",
@@ -164,7 +165,6 @@
    "columns": 3,
    "fieldname": "warehouse",
    "fieldtype": "Link",
-   "in_list_view": 1,
    "label": "Target Warehouse",
    "oldfieldname": "warehouse",
    "oldfieldtype": "Link",
@@ -461,7 +461,7 @@
  "index_web_pages_for_search": 1,
  "istable": 1,
  "links": [],
- "modified": "2023-05-07 20:23:31.250252",
+ "modified": "2024-01-04 16:10:36.866727",
  "modified_by": "Administrator",
  "module": "Stock",
  "name": "Material Request Item",
